Clean up unused components
