+++ src/domain/space/about/SpaceAboutPage.tsx
@@ -1,8 +1,6 @@
-import { useState } from 'react';
 import { useSpace } from '../context/useSpace';
 import { useBackWithDefaultUrl } from '@/core/routing/useBackToPath';
 import SpaceAboutDialog from '@/domain/space/about/SpaceAboutDialog';
-import SpaceContributorsDialog from '@/domain/community/community/ContributorsDialog/SpaceContributorsDialog';
 import { StorageConfigContextProvider } from '@/domain/storage/StorageBucket/StorageConfigContext';
 import { useSpaceAboutDetailsQuery } from '@/core/apollo/generated/apollo-hooks';
 import { Box } from '@mui/material';
@@ -17,8 +15,6 @@ const SpaceAboutPage = () => {
   const spaceDetails = data?.lookup.space;
 
   const backToParentPage = useBackWithDefaultUrl(permissions.canRead ? space.about.profile.url : undefined);
-
-  const [isContributorsDialogOpen, setIsContributorsDialogOpen] = useState(false);
 
   return (
     <>
@@ -35,11 +31,6 @@ const SpaceAboutPage = () => {
             hasEditPrivilege={permissions.canUpdate}
           />
         )}
-        <SpaceContributorsDialog
-          open={isContributorsDialogOpen}
-          onClose={() => setIsContributorsDialogOpen(false)}
-          roleSetId={spaceDetails?.about.membership.roleSetID}
-        />
       </StorageConfigContextProvider>
     </>
   );
